requirements.txt
asgiref==3.3.1
Django==3.1.7
django-stdimage==5.2.0
django-summernote==0.8.11.6
Pillow==8.1.2
progressbar2==3.53.1
python-utils==2.5.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
